README.md
I need to read in my ebook one programation book that is in photo pdf. I did this little script on the run to convert from pdf to ebook. The aspose module work but I need license to get all the pages, but finally my ebook accept .txt so it works.